requirements.txt
streamlit==1.36.0
feedparser==6.0.11
requests==2.32.3
python-dateutil==2.9.0.post0
pytz==2024.1
beautifulsoup4==4.12.3
readability-lxml==0.8.1
trafilatura==1.9.0
sqlite-utils==3.37
sentence-transformers==3.0.1
chromadb==0.5.5
scikit-learn==1.5.1
pydantic==2.8.2
pyyaml==6.0.1
tqdm==4.66.4
tenacity==8.4.2
ragas==0.1.17
openai==0.28
